Created LottoCoupon class instead of having it in Lotto Servlet
